app/requirements.txt
flask
sqlalchemy
pymysql
mysql-connector-python
flask-sqlalchemy
requests
nltk
numpy
pandas
python-dotenv
